ResearchAgent/requirements.txt
langchain
langchain-groq
langchain-community
python-dotenv
pydantic
wikipedia 
langchain-core
duckduckgo-search